requirements.txt
discord.py
asyncio
discord.ext
discord.ext.commands